Write a simple test that tests the ping handler.
