+++ .vscode/launch.json
@@ -4,6 +4,24 @@
     // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
     "version": "0.2.0",
     "configurations": [
+    {
+        "type": "node",
+        "request": "launch",
+        "name": "Mocha Tests",
+        "program": "${workspaceFolder}/node_modules/mocha/bin/_mocha",
+        "args": [
+            "-u",
+            "bdd",
+            "--timeout",
+            "999999",
+            "--colors",
+            "${workspaceFolder}/tests"
+        ],
+        "internalConsoleOptions": "openOnSessionStart",
+        "skipFiles": [
+            "<node_internals>/**"
+        ]
+    },
         {
             "type": "node",
             "request": "launch",
@@ -27,7 +45,22 @@
                 "<node_internals>/**"
             ],
             "program": "${workspaceFolder}/bin/www"
-        }
+        },
+        {
+            "type": "node",
+            "request": "launch",
+            "name": "Debug Tests via NPM",
+            
+            "runtimeExecutable": "npm",
+            "runtimeArgs": [
+                "run-script",
+                "test"
+            ],
+            "port": 9229,
+            "skipFiles": [
+                "<node_internals>/**"
+            ]
+        },        
         /*,
         {
             "type": "node",
