+++ package.json
@@ -4,7 +4,9 @@
   "private": true,
   "scripts": {
     "start": "node ./bin/www",
-    "debug": "nodemon --inspect ./bin/www"
+    "debug": "nodemon --inspect ./bin/www",
+    "coverage": "nyc mocha --timeout=3000",
+    "test": "mocha ./tests"
   },
   "dependencies": {
     "cookie-parser": "~1.4.4",
@@ -14,5 +16,9 @@
     "http-errors": "~1.6.3",
     "morgan": "~1.9.1",
     "stylus": "0.54.5"
+  },
+  "devDependencies": {
+    "chai": "^4.2.0",
+    "mocha": "^6.2.2"
   }
 }
